refactor(api): stock 관련 API 라우트 리팩토링
- 새로운 피처 아키텍처에 맞춰 API 라우트 업데이트
- stock 관련 모든 API 엔드포인트 수정
- posts API에서 stock 의존성 업데이트
