+++ apps/web/src/app/api/stocks/[symbol]/route.ts
@@ -1,4 +1,4 @@
-import { StockRepository } from '@/entities/stock/repository';
+import { StockRepository } from '@/features/stock/server/repository';
 import { NextRequest, NextResponse } from 'next/server';
 
 export const dynamic = 'force-dynamic';
