+++ apps/web/src/app/api/stocks/price/overseas/route.ts
@@ -1,4 +1,4 @@
-import { StockService } from '@/features/stock/service/stock.server';
+import { StockService } from '@/features/stock/server/service/stock.service';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(req: NextRequest) {
