+++ apps/web/src/app/api/stocks/route.ts
@@ -1,6 +1,6 @@
 import type { Stock } from '@prisma/client';
 import { NextRequest, NextResponse } from 'next/server';
-import { StockRepository } from '@/entities/stock/repository';
+import { StockRepository } from '@/features/stock/server/repository';
 
 export async function GET(request: NextRequest) {
     try {
